requirements.txt
pycocotools
opencv-python
matplotlib
chumpy
trimesh
tqdm
yacs
torchgeometry
pyrender
smplx
pyopengl==3.1.4